requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
anthropic>=0.40.0
python-dotenv>=1.0.0
streamlit>=1.30.0
